controller/apps/backend/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
dbus-python==1.2.16
docker==5.0.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-networkmanager==2.2
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
urllib3==1.26.5
websocket-client==1.1.0
Werkzeug==2.0.1